Lighting and Level Design Updatea
Added windows around in the walls on 2 walls  on each side, spot lighting under new ceiling lights one of them is yellow over the cheese which i thought was an alright touch, made the walls more metallic so it looks more like a kitchen imo, and the lighting the reason shadows didnt work is because we have to use point lighting instead of directional, improved it and added a more yellowwish tone to make it look like its morning. (i like the aesthetic u can change it if u want im dumb), added more clutter and extended the table on the left side so it is not standing in the corner anymore with all that empty space, dialed down the cat speed to 3.2 for first cat and 3.5 for second cat since there are more objects for it to get stuck on, moved some things around a bit and added more clutter on the floor and tables,  tinked with the level design a bit just adding more stuff, i just finished all this lmao i am a slow ass worker holy shit. i have class at 10:30 LOL. once again i apologize for not being there for the team last night hopefully some of this work makes up for it a bit